Broke a bunch of stuff, then fixed it lol.
Reworked the timeline threading logic and added support for home feed preferences-based filtering.
The latter parallels pretty closely the official app's logic by necessity and similarly could support similar tuning for custom feeds. The former differs a fair bit.

My method is more network-intensive in exchange for actually pulling long threads of self-replies together and having them bump the timeline. Currently there's no time limit, so you can go back a while, so long as it's within ten hops up, but I implemented a show/collapse toggle that appears when threads are longer to make them not dominate the timeline.

Potentially considering doing the network calls more asynchronously and updating the UI somehow to speed initial timeline rendering and immediate scrollability, but don't want the timeline to jump around either. Might be able to do something with itemsIndexed in LazyColumn to help that?

Another way to avoid making so many network requests in quick succession would be to batch up the URIs (are they all present in the response data or..., it's hard to figure out looking at the debugger or longs, need to write a custom toString()-type function for FeedViewPost or BskyPost that doesn't log all the post content) and request them all at once, then sort out the threads they belong to. The delay is afaik per request, so one larger request would be quicker (though maybe could try and make the requests simultaneously?), in exchange for a bit more processing on the client-side.