Daily commit
Fixed Poll.blue crashing the app.

Fixed thread comparator crashing the app (blocked posts don't have times, so it went to the hashcode, which is an Int, not an Instant. Converted time to unix epoch seconds, then cast hashcode to a Long as well.

Added basic navigation and some placeholder text to the unfinished screens so they're not dead ends.

Stopped the local data model from discarding the image aspect ratio data present in the actual lexicon (why, Christian?). That's useful to have to estimate the size of the view in advance for rendering.